requirements_test.txt
# Testing and Database Requirements
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
